[#117] loglevel default logger
