+++ lib/config.js
@@ -10,6 +10,8 @@ const path = require('path')
 const fs = require('fs')
 const defaultConfig = require('./pinpoint-config-default')
 const log = require('./utils/logger')
+const { setLog } = require('./supports')
+const { makeLogLevelLog } = require('./utils/log/log-level-logger')
 
 const valueOfString = (envName) => {
   return () => {
@@ -142,6 +144,7 @@ const init = (initOptions = {}) => {
     readConfigJson(initOptions))
 
   log.init(agentConfig.logLevel)
+  setLog(makeLogLevelLog(agentConfig.logLevel))
 
   Object.entries(REQUIRE_CONFIG).forEach(([propertyName, description]) => {
     if (agentConfig.enable && !agentConfig[propertyName]) {
@@ -247,15 +250,10 @@ function hasDockerCGroup() {
   }
 }
 
-function needsLoadConfig() {
-  return !agentConfig
-}
-
 module.exports = {
   getConfig,
   clear,
   readConfigJson,
   readRootConfigFile,
-  getMainModulePath,
-  needsLoadConfig
+  getMainModulePath
 }
