+++ lib/supports.js
@@ -7,8 +7,6 @@
 'use strict'
 
 const Logger = require('./utils/log/logger2')
-const { getConfig, needsLoadConfig } = require('./config')
-const defaultLog = require('loglevel').getLogger('default_logger')
 
 let log = undefined
 module.exports = {
@@ -17,8 +15,8 @@ module.exports = {
          return log
       }
 
-      if (needsLoadConfig()) {
-         return new Logger.NoneBuilder({
+      if (!log) {
+         log = new Logger.NoneBuilder({
             output: console,
             debug: function (message) {
                this.output.debug(message)
@@ -34,29 +32,9 @@ module.exports = {
             }
          }).build()
       }
-
-      const config = getConfig()
-      if (!log) {
-         defaultLog.setLevel(Logger.logTypeNameOf(config.logLevel))
-         log = Logger.makeBuilder(config.logLevel, {
-            log: defaultLog,
-            debug: function (message) {
-               this.log.debug(message)
-            },
-            info: function (message) {
-               this.log.info(message)
-            },
-            warn: function (message) {
-               this.log.warn(message)
-            },
-            error: function (message) {
-               this.log.error(message)
-            }
-         }).build()
-      }
       return log
    },
-   clearLog: function () {
-      log = undefined
+   setLog: function(logger) {
+      log = logger
    }
 }
